fix(runtime): preserve cancellation outcome classification
