+++ pkg/services/factory_runtime/internal/services/orchestration/engine/engine.go
@@ -240,51 +240,6 @@ func (e *FactoryEngine) retireCompletedDispatches(results []workerexecution.Work
 	}
 }
 
-func completedDispatchReasonFromResult(result workerexecution.WorkResult) string {
-	switch result.Outcome {
-	case workerexecution.OutcomeCanceled:
-		if result.Cancellation != nil {
-			return string(result.Cancellation.Reason)
-		}
-		return string(workerexecution.DispatchCancellationReasonCanceled)
-	case workerexecution.OutcomeFailed:
-		return result.Error
-	case workerexecution.OutcomeContinue:
-		return result.Feedback
-	case workerexecution.OutcomeRejected:
-		return result.Feedback
-	default:
-		return ""
-	}
-}
-
-func workResultForCompletedDispatch(result workerexecution.WorkResult, completed interfaces.CompletedDispatch) workerexecution.WorkResult {
-	result.Outcome = completed.Outcome
-	result.Cancellation = completed.Cancellation.Clone()
-	if completed.Outcome == workerexecution.OutcomeCanceled {
-		result.Error = completed.Reason
-		result.FailureDetail = nil
-		result.FailureMetadata = nil
-		result.RecordedOutputWork = nil
-		result.Output = ""
-		result.StructuredResult = nil
-		result.StructuredResultPresent = false
-	}
-	result.SelectedClassificationLabel = completed.SelectedClassificationLabel
-	if completed.FailureDetail != nil {
-		result.FailureDetail = workerexecution.CloneFailureDetail(completed.FailureDetail)
-	}
-	switch completed.Outcome {
-	case workerexecution.OutcomeFailed:
-		result.Error = completed.Reason
-	case workerexecution.OutcomeContinue:
-		result.Feedback = completed.Reason
-	case workerexecution.OutcomeRejected:
-		result.Feedback = completed.Reason
-	}
-	return result
-}
-
 // NotifyResult wakes the engine after a WorkResult is enqueued so the engine
 // ticks and routes the result. Non-blocking: drops if the buffer is full.
 func (e *FactoryEngine) NotifyResult() {
