+++ pkg/services/factory_runtime/internal/services/orchestration/runtime/work_output_materialization.go
@@ -2,14 +2,12 @@ package runtime
 
 import (
 	"context"
-	"errors"
 	"fmt"
 	pathpkg "path"
 	"path/filepath"
 	"sort"
 	"strings"
 
-	"github.com/portpowered/infinite-you/pkg/platform/jsonvalue"
 	interfaces "github.com/portpowered/infinite-you/pkg/services/factory_definitions"
 	"github.com/portpowered/infinite-you/pkg/services/factory_runtime/internal/services/orchestration/state"
 	"github.com/portpowered/infinite-you/pkg/services/providers"
@@ -277,106 +275,6 @@ func isASCIIAlphaRuntime(value byte) bool {
 	return (value >= 'a' && value <= 'z') || (value >= 'A' && value <= 'Z')
 }
 
-func workstationDispatchResultFromExecute(
-	request workers.WorkstationDispatchRequest,
-	result workers.ExecuteResult,
-	executeErr error,
-) (workers.WorkstationDispatchResult, error) {
-	dispatch := request.Execution.Dispatch
-	proposedOutput := result.Output.Clone()
-	workResult := workers.WorkResult{
-		DispatchID:                  dispatch.DispatchID,
-		TransitionID:                dispatch.TransitionID,
-		Outcome:                     workers.OutcomeAccepted,
-		Cancellation:                result.Cancellation.Clone(),
-		Output:                      primaryOutputText(result.Output.Primary),
-		StructuredResult:            jsonvalue.Clone(result.StructuredResult),
-		StructuredResultPresent:     jsonvalue.Present(result.StructuredResult, result.StructuredResultPresent),
-		ArtifactVerification:        result.ArtifactVerification.Clone(),
-		Feedback:                    result.Output.Feedback,
-		SelectedClassificationLabel: result.Output.Classification,
-		Metrics: workers.WorkMetrics{
-			Duration:   result.Metrics.Duration,
-			Cost:       result.Metrics.Cost,
-			RetryCount: result.Metrics.RetryCount,
-		},
-		Continuation: result.Continuation,
-		Diagnostics:  result.Diagnostics.ToWorkDiagnostics(),
-	}
-	terminal := workers.WorkstationDispatchTerminalOutcomeCompleted
-	switch result.Outcome {
-	case workers.ExecutionOutcomeContinue:
-		workResult.Outcome = workers.OutcomeContinue
-	case workers.ExecutionOutcomeRejected:
-		workResult.Outcome = workers.OutcomeRejected
-	case workers.ExecutionOutcomeFailed:
-		workResult.Outcome = workers.OutcomeFailed
-		terminal = workers.WorkstationDispatchTerminalOutcomeFailed
-	case workers.ExecutionOutcomeCanceled:
-		workResult.Outcome = workers.OutcomeCanceled
-		terminal = workers.WorkstationDispatchTerminalOutcomeCanceled
-		if workResult.Cancellation == nil {
-			workResult.Cancellation = workers.NewDispatchCancellation(workers.DispatchCancellationReasonCanceled)
-		}
-	default:
-		if result.Outcome != workers.ExecutionOutcomeAccepted {
-			workResult.Outcome = workers.OutcomeFailed
-			terminal = workers.WorkstationDispatchTerminalOutcomeFailed
-		}
-	}
-	if result.Failure != nil {
-		workResult.Error = strings.TrimSpace(result.Failure.Message)
-		workResult.FailureDetail = workFailureDetail(result.Failure, workResult.Error)
-		if shouldPropagateFailureMetadata(request, result.Failure) {
-			workResult.FailureMetadata = &workers.WorkFailureMetadata{
-				Family: result.Failure.Family,
-				Type:   result.Failure.Type,
-			}
-		}
-	}
-	if workResult.Cancellation != nil {
-		workResult.Outcome = workers.OutcomeCanceled
-		terminal = workers.WorkstationDispatchTerminalOutcomeCanceled
-	}
-	if errors.Is(executeErr, context.Canceled) || errors.Is(executeErr, workers.ErrWorkstationDispatchCanceled) {
-		workResult.Outcome = workers.OutcomeCanceled
-		terminal = workers.WorkstationDispatchTerminalOutcomeCanceled
-		if workResult.Cancellation == nil {
-			workResult.Cancellation = workers.NewDispatchCancellation(workers.DispatchCancellationReasonCanceled)
-		}
-	}
-	if executeErr != nil && terminal != workers.WorkstationDispatchTerminalOutcomeCanceled {
-		terminal = workers.WorkstationDispatchTerminalOutcomeFailed
-		workResult.Outcome = workers.OutcomeFailed
-		if strings.TrimSpace(workResult.Error) == "" {
-			workResult.Error = executeErr.Error()
-		}
-	}
-	if terminal == workers.WorkstationDispatchTerminalOutcomeCanceled && strings.TrimSpace(workResult.Error) == "" {
-		workResult.Error = workers.ErrWorkstationDispatchCanceled.Error()
-	}
-	if terminal == workers.WorkstationDispatchTerminalOutcomeCanceled {
-		workResult.Outcome = workers.OutcomeCanceled
-		workResult.Output = ""
-		workResult.StructuredResult = nil
-		workResult.StructuredResultPresent = false
-		workResult.Continuation = nil
-		workResult.FailureDetail = nil
-		workResult.FailureMetadata = nil
-		proposedOutput = workers.ProposedOutput{}
-	}
-	reconciliationReason := processGoneDispatchResult(&workResult, &terminal, executeErr)
-	return workers.WorkstationDispatchResult{
-		DispatchID:           dispatch.DispatchID,
-		WorkstationName:      request.WorkstationName,
-		TerminalOutcome:      terminal,
-		ReconciliationReason: reconciliationReason,
-		Cancellation:         workResult.Cancellation.Clone(),
-		Result:               workResult,
-		ProposedOutput:       &proposedOutput,
-	}, executeErr
-}
-
 func workFailureDetail(failure *workers.ExecutionFailure, message string) *workers.FailureDetail {
 	detail := workers.CloneFailureDetail(failure.Detail)
 	if detail != nil || message == "" {
@@ -526,7 +424,7 @@ func applyMaterializedWorkerOutput(
 	if result.Cancellation != nil || result.Outcome == workerexecution.OutcomeCanceled {
 		result.Outcome = workerexecution.OutcomeCanceled
 		if result.Cancellation == nil {
-			result.Cancellation = workerexecution.NewDispatchCancellation(workerexecution.DispatchCancellationReasonCanceled)
+			result.Cancellation = &workerexecution.DispatchCancellation{Reason: workerexecution.DispatchCancellationReasonCanceled}
 		}
 		result.Output = ""
 		result.StructuredResult = nil
